requirements.txt
python-telegram-bot==13.3
python-dotenv==0.19.1
pandas==1.2.1